refactor: 패키지 분리
패키지를 도메인에 따라 분류한다.
