+++ src/main/java/calculator/StringCalculator.java
@@ -1,7 +1,14 @@
 package calculator;
 
+import calculator.adder.IntegerAdder;
 import calculator.io.ConsoleInputHandler;
 import calculator.io.ConsoleOutputHandler;
+import calculator.util.DelimiterExtractor;
+import calculator.util.NumberConverter;
+import calculator.util.RegexGenerator;
+import calculator.util.Splitter;
+import delimiter.Delimiter;
+import delimiter.Delimiters;
 import java.util.Arrays;
 import java.util.List;
