+++ src/main/java/calculator/util/DelimiterExtractor.java
@@ -1,5 +1,7 @@
-package calculator;
+package calculator.util;
 
+import delimiter.Delimiter;
+import delimiter.Delimiters;
 import java.util.ArrayList;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
