+++ src/main/java/calculator/util/RegexGenerator.java
@@ -1,5 +1,7 @@
-package calculator;
+package calculator.util;
 
+import delimiter.Delimiter;
+import delimiter.Delimiters;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
