+++ src/main/java/delimiter/Delimiters.java
@@ -1,4 +1,4 @@
-package calculator;
+package delimiter;
 
 import java.util.ArrayList;
 import java.util.Collections;
